requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
google-api-python-client==2.84.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
twilio==8.0.0
openai==0.27.0

